pyproject.toml
[project]
name = "ddr-report-generator"
version = "0.1.0"
description = "AI-powered Detailed Diagnosis Report (DDR) generator from inspection PDFs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.4",
    "langchain-huggingface>=0.1.0",
    "langchain-cohere>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pymupdf>=1.24.0",
    "reportlab>=4.2.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "cohere>=5.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
generate-ddr = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
